+++ src/sovereign/views/healthchecks.py
@@ -1,33 +1,34 @@
 import random
-from quart import Blueprint
+from fastapi.routing import APIRouter
+from starlette.responses import PlainTextResponse
 from sovereign import XDS_TEMPLATES, __versionstr__
 from sovereign import discovery
 from sovereign.sources import match_node
 from sovereign.utils.mock import mock_discovery_request
 
 
-blueprint = Blueprint('healthchecks', __name__)
+router = APIRouter()
 
 
-@blueprint.route('/healthcheck')
-def health_check():
-    return 'OK'
+@router.get('/healthcheck', summary='Does the server respond')
+async def health_check():
+    return PlainTextResponse('OK')
 
 
-@blueprint.route('/deepcheck')
-def deep_check():
+@router.get('/deepcheck', summary='Can the server render a random template')
+async def deep_check():
     template = random.choice(
         list(XDS_TEMPLATES['default'].keys())
     )
-    discovery.response(
+    await discovery.response(
         mock_discovery_request(),
         xds=template,
         debug=True
     )
     match_node(request=mock_discovery_request())
-    return 'OK'
+    return PlainTextResponse('OK')
 
 
-@blueprint.route('/version')
-def version_check():
-    return f'Sovereign {__versionstr__}'
+@router.get('/version', summary='Display the current version of Sovereign')
+async def version_check():
+    return PlainTextResponse(f'Sovereign {__versionstr__}')